refactor: winningTicket과 Ticket을 분리하지 않도록 수정
